pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rvm-eupe"
version = "0.1.0"
description = "Recurrent Video MAE with EUPE Backbone"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "torchvision",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "einops>=0.7",
    "wandb",
    "webdataset>=0.2.86",
    "timm>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rvm_eupe*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
